python-service/requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
pydantic==2.11.9
newspaper3k==0.2.8
ddgs==9.6.0
lxml_html_clean==0.4.2